Added Auth logic for Store Managers

+++ source/controllers/store.controller.ts
@@ -5,6 +5,7 @@ import { RequestHelper } from "../helpers/request.helpers";
 import { ResponseHelper } from "../helpers/response.helper";
 import { StoreService } from "../services/store.service";
 import { NON_EXISTING_ID } from "../constants";
+import { Role } from "../enums";
 
 const errorService: ErrorService = new ErrorService();
 const storeService: StoreService = new StoreService(errorService);
@@ -61,10 +62,21 @@ const updateStoreById = async (req: Request, res: Response, next: NextFunction) 
     const numericParamOrError: number | systemError = RequestHelper.parseNumericInput(errorService, req.params.id);
 
     if (typeof numericParamOrError === "number") {
-        if (numericParamOrError > 0) {
+        const storeId = numericParamOrError; // just alias for better reading
+        const userId = (req as AuthenticatedRequest).userData.userId;
+        const userRoles: Role[] = (req as AuthenticatedRequest).userData.rolesId;
+
+        const isUserHasAccess: boolean = await RequestHelper.isUserHasAccessToStore(
+            errorService, userId, userRoles, storeId)
+        
+        if (!isUserHasAccess) {
+                return res.sendStatus(401);
+        }
+
+        if (storeId > 0) {
             const body: storeType = req.body;
             const store = {
-                id: numericParamOrError,
+                id: storeId,
                 name: body.name,
                 address: body.address,
                 openDate: body.openDate,
@@ -101,6 +113,13 @@ async function addNewStore(req: Request, res: Response, next: NextFunction) {
         scale: body.scale
     };
 
+    // Adding store is prohibited for SM
+    const userRoles: Role[] = (req as AuthenticatedRequest).userData.rolesId;
+    const isUserStoreManager = userRoles.indexOf(Role.StoreManager) > -1;
+    if (isUserStoreManager) {
+        return res.sendStatus(401);
+    }
+
     storeService.addNewStore(inputStore, (req as AuthenticatedRequest).userData.userId)
         .then((result: storeType) => {
             return res.status(200).json(result);
@@ -114,6 +133,17 @@ async function deleteStore(req: Request, res: Response, next: NextFunction) {
     const numericParamOrError: number | systemError = RequestHelper.parseNumericInput(errorService, req.params.id);
 
     if (typeof numericParamOrError === "number") {
+        const storeId = numericParamOrError; // just alias for better reading
+        const userId = (req as AuthenticatedRequest).userData.userId;
+        const userRoles: Role[] = (req as AuthenticatedRequest).userData.rolesId;
+
+        const isUserHasAccess: boolean = await RequestHelper.isUserHasAccessToStore(
+            errorService, userId, userRoles, storeId)
+        
+        if (!isUserHasAccess) {
+                return res.sendStatus(401);
+        }
+        
         if (numericParamOrError > 0) {
             storeService.deleteStore(numericParamOrError, (req as AuthenticatedRequest).userData.userId)
                 .then(() => {
